Feature users

+++ app/Http/Controllers/PillarController.php
@@ -8,6 +8,7 @@ use App\Models\Objective;
 use App\Models\Subobjective;
 use App\Models\Routine;
 
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -17,7 +18,10 @@ class PillarController extends Controller
 {
     public function index()
     {
-        $pillars = Pillar::with('subpillars.objectives.subobjectives')->get();
+        // Récupère seulement les piliers de l'utilisateur connecté
+        $pillars = Pillar::with('subpillars.objectives.subobjectives')
+                         ->where('user_id', Auth::id()) // Filtre par user_id
+                         ->get();
 
         return Inertia::render('Custom/GoalPage', [
             'pillars' => $pillars,
@@ -30,34 +34,40 @@ class PillarController extends Controller
             'description' => ['required'],
         ])->validate();
 
+        $valid_data['user_id'] = auth()->id();
+
         $pillar = Pillar::create($valid_data);
+
+        return redirect()->route('pillars.index')->with('success', 'Pillar created successfully.');
     }
+
 
     public function update(Request $request, Pillar $pillar)
     {
-    // Validation des données reçues
-    $valid_data = Validator::make($request->all(), [
-        'name' => ['required', 'string', 'max:255'],
-        'description' => ['required', 'string'],
-    ])->validate();
+        // Assurez-vous que le pilier appartient à l'utilisateur connecté
+        if ($pillar->user_id != Auth::id()) {
+            return redirect()->back()->with('error', 'Unauthorized action.');
+        }
 
-    // Mise à jour du pilier
-    $pillar->update($valid_data);
+        $valid_data = Validator::make($request->all(), [
+            'name' => ['required', 'string', 'max:255'],
+            'description' => ['required', 'string'],
+        ])->validate();
 
-    // Redirection ou retour d'une réponse, selon votre application
-    // Par exemple, redirigez vers la page d'index avec un message de succès :
-    return redirect()->route('pillars.index')->with('success', 'Pillar updated successfully.');
+        $pillar->update($valid_data);
+
+        return redirect()->route('pillars.index')->with('success', 'Pillar updated successfully.');
     }
 
     public function destroy(Pillar $pillar)
     {
-    // Assurez-vous d'ajouter des contrôles de sécurité ici, comme vérifier les permissions de l'utilisateur
+        // Vérifie que le pilier appartient à l'utilisateur connecté avant de supprimer
+        if ($pillar->user_id != Auth::id()) {
+            return redirect()->back()->with('error', 'Unauthorized action.');
+        }
 
-    $pillar->delete();
+        $pillar->delete();
 
-    // Rediriger l'utilisateur avec un message de succès
-    return redirect()->back()->with('success', 'Pillar deleted successfully.');
+        return redirect()->back()->with('success', 'Pillar deleted successfully.');
     }
-
-
 }
